% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherence.R
\name{coherence}
\alias{coherence}
\title{Title}
\usage{
coherence(Var, cond, com, fichier)
}
\arguments{
\item{Var}{variable}

\item{cond}{condition}

\item{com}{commentaire}

\item{fichier}{dataframe}
}
\description{
Title
}
