% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherence3.R
\name{coherence3}
\alias{coherence3}
\title{Title}
\usage{
coherence3(Var, cond, com, fichier)
}
\arguments{
\item{Var}{variable à controler}

\item{cond}{condition à vérifier pour la variable}

\item{com}{commentaire}

\item{fichier}{dataframe où se trouve la variable}
}
\description{
Title
}
