% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setFormat.R
\name{setFormat}
\alias{setFormat}
\title{Title}
\usage{
setFormat(df, dico)
}
\arguments{
\item{df}{dataframe à recoder(pour les variables catégorielles)}

\item{dico}{dictionnaire de données(dhm)}
}
\description{
Title
}
