% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whereIsMyVar.R
\name{whereIsMyVar}
\alias{whereIsMyVar}
\title{Title}
\usage{
whereIsMyVar(z)
}
\arguments{
\item{z}{variable a chercher dans les dataframes}
}
\value{
le nom du dataframe est affiché à l'écran et la variable moduel_var contient le nom du dataframe
}
\description{
Title
}
