% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeDataFrame.R
\name{writeDataFrame}
\alias{writeDataFrame}
\title{Title}
\usage{
writeDataFrame(df, sheetName, fileName)
}
\arguments{
\item{df}{dataframe à exporter en fichier Excel}

\item{sheetName}{le nom de la feuille Excel}

\item{fileName}{le nom du fichier Excel}
}
\description{
Title
}
