% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeQuery2.R
\name{writeQuery2}
\alias{writeQuery2}
\title{Title}
\usage{
writeQuery2(data, cheminQuery, etude)
}
\arguments{
\item{data}{dataframe test}

\item{cheminQuery}{chemin des queries}

\item{etude}{nom de l'étude}
}
\description{
Title
}
